New update. [1] Navigation moved to App.tsx from Header.tsx for more controll. [2] HeaderProps interface created. [3] Created some router in Main.tsx.
