+++ src/sections/app/main/Info.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+function Info() {
+    return (
+        <div>
+            <h1>Hello world</h1>
+        </div>
+    );
+}
+
+export default Info;
